==> data analysis/1a_Exp1_Plot.Rmd <==

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

set.seed(444) #Set up random seed for reproducibility

options(scipen = 999) #Turn off scientific expressions

library(tidyverse)
library(qdapTools)
library(Rmisc)

library(lme4)
library(lmerTest)
library(buildmer)

library(phia)

library(r2glmm)
library(effsize)
library(performance)

rm(list=ls()) #Clear environment

#load("Exp1.RData")
```

#Import & Normalization

```{r}
data1 <- read.csv("//wcs-cifs/wc/smng/experiments/simonMultisyllable/acousticdata/R_Zyy/Manuscript/F1_Exp1_S1.csv")

data1 <- unique(data1)

#Add Chunk info

var1 <- length(unique(data1$Word)) * 10

data1$Chunk <- ceiling(data1$Trial/var1)

#Obtain Baseline means

temp1 <- aggregate(data1$meanF1_mel, 
                   list(data1$Direction, data1$Chunk, data1$Speaker), 
                   FUN = mean) 

colnames(temp1) <- c("Direction", "Chunk", "Speaker", "meanF1")

#Limit results to the last 10 trials of Baseline 

temp2 <- data.frame(table(data1$Chunk, data1$Phase))
temp2 <- temp2[which(temp2$Freq != 0 & temp2$Var2 == "Baseline"),]
var1 <- max(as.numeric(as.character(temp2$Var1)))


temp1 <- temp1[which(temp1$Chunk == var1), ]

temp1$Tracer <- paste(temp1$Speaker, temp1$Direction, sep = "_")
data1$Tracer <- paste(data1$Speaker, data1$Direction, sep = "_")

#Pairing data from temp1 

data1$F1_Baseline <- lookup(terms = data1$Tracer, 
                            key.match = temp1$Tracer,
                            key.reassign = temp1$meanF1)

#Calculate F1 change 

data1$F1_Change <- data1$meanF1_mel - data1$F1_Baseline
```

```{r}
#Add frequency

data1$Lexical <- ifelse(data1$Word == "head", log10(0.0385950698/100), 
                        ifelse(data1$Word == "bed", log(0.0127859079/100), log(0.0000305076/100)))

table(data1$Lexical, data1$Word)

#Add Block info

var1 <- max(data1$Trial)
var2 <- max(data1$Trial)/length(unique(data1$Word))

temp1 <- seq(from = 1, to = var1, by = 1)

temp2 <- rep(1:var2, each = length(unique(data1$Word)), length.out = var1)

temp1 <- data.frame(temp1, temp2)

colnames(temp1) <- c("Trial", "Block")

#Pair
data1$Block <- lookup(terms = data1$Trial,
                      key.match = temp1$Trial,
                      key.reassign = temp1$Block)

save.image("Exp1.RData")
```

#Main, results
```{r}
#Set up bars and labels

bar_1 = 35
bar_2 = 65
bar_3 = 155

var_Baseline = bar_1 / 2
var_Ramp = (bar_1 + bar_2) /2 
var_Hold = (bar_2 + bar_3) /2 
var_Washout = bar_3 + 15
```

```{r}
#Plot grand average

data1$Direction <- factor(data1$Direction, levels = c("Upshift", 
                                                      "Downshift", 
                                                      "Unshifted"))

temp1 <- summarySEwithin(data = data1, 
                         measurevar = "F1_Change",
                         withinvars = c("Direction", "Chunk"), 
                         idvar = "Speaker",
                         na.rm = TRUE)

temp1$Chunk <- as.numeric(temp1$Chunk) * 10
```

```{r}
ggplot(data = temp1, aes(x = Chunk, y = F1_Change, 
                               color = Direction)) +
  geom_point() + 
  geom_line() + 
  geom_errorbar(aes(ymin = F1_Change - se, ymax = F1_Change + se), 
                width = 4,
                position = position_dodge(0)) + 
  theme_classic() + 
  ylab("F1 change (mel)") + 
  geom_hline(yintercept = 0, linetype = "solid", 
             color = "black", linewidth = 0.5) + 
  geom_vline(xintercept = bar_1, linetype = "twodash", 
             color = "black", linewidth = 0.7) + 
  geom_vline(xintercept = bar_2, linetype = "twodash", 
             color = "black", linewidth = 0.7) + 
  geom_vline(xintercept = bar_3, linetype = "twodash", 
             color = "black", linewidth = 0.7) + 
  annotate("rect", xmin = 155, xmax = 165, ymin = -35, ymax = 45,
           alpha = 0.2,fill = "yellow") + 
  annotate("rect", xmin = 125, xmax = 135, ymin = -35, ymax = 45,
           alpha = 0.2,fill = "yellow") + 
  annotate(geom = "text", x = var_Baseline, y = max(temp1$F1_Change), 
           label = "Baseline",color = "black", size = 5) +  
  annotate(geom = "text", x = var_Ramp, y = max(temp1$F1_Change), 
           label = "Ramp", color = "black", size = 5) + 
  annotate(geom = "text", x = var_Hold, y =  max(temp1$F1_Change), 
           label = "Hold", color = "black", size = 5) + 
  annotate(geom = "text", x = var_Washout, y =  max(temp1$F1_Change), 
           label = "Washout", color = "black", size = 5) +
  theme(legend.position = "none", 
        axis.text = element_text(size = 15), 
        axis.title.y = element_text(size = 18), 
        axis.title.x = element_blank(),
        axis.text.x = element_blank()) + 
  scale_y_continuous(expand = c(0, 0), limits = c(-35, 46), 
                     breaks = seq(-40, 50, by = 10)) +
  scale_x_continuous(expand = c(0, 0), limits = c(1, 183), 
                     breaks = c(10, 30, 40, 60, 
                                70, 150, 160, 180)) +
  scale_color_manual(values = c("red", "blue", "gray")) 

ggsave(filename = "H:/Shared drives/SMNG/Experiments/simonMultisyllable/Manuscript/PlotGen/1B.png",
       plot = last_plot(), 
       dpi = 600,
       width = 9,
       height = 4)
```


#Main, perturbations 

```{r}
#Numbers
Chunks <- c(1, 33.99, 65.999, 150, 155.999, 156, 185)

Shifts <- c("Unshifted", "Upshift", "Downshift")

temp1 <- expand.grid(Chunks, Shifts)

colnames(temp1) <- c("Chunk", "Direction")

temp1$Mags <- c(0, 0, 0, 0, 0, 0, 0, 
                2.0, 2.0, 125, 125, 125, 2.0, 2.0,
                -2.0, -2.0, -125, -125, -125, -2.0, -2.0)

temp1$Direction <- factor(temp1$Direction, levels = c("Upshift", 
                                                      "Downshift",
                                                      "Unshifted"))
```

```{r}
#Plot
ggplot(data = temp1, 
             aes(x = Chunk, y = Mags, color = Direction)) +
  geom_line(linewidth = 1) + 
  theme_classic() + 
  ylab("F1 pert (mel)") + 
  theme(legend.position = "none", 
        axis.title.x = element_blank(),
        axis.text.x = element_blank(),
        axis.ticks.x = element_blank()) + 
  scale_x_continuous(expand = c(0, 0), limits = c(1, 185), 
                     breaks = c(10, 30, 40, 60, 
                                70, 150, 160, 180)) +
  scale_y_continuous(expand = c(0, 0), limits = c(-130, 130), 
                     breaks = c(-125, 0, 125)) +
  scale_color_manual(values = c("red", "blue", "gray"))

ggsave(filename = "H:/Shared drives/SMNG/Experiments/simonMultisyllable/Manuscript/PlotGen/1A.png",
       plot = last_plot(), 
       dpi = 600,
       width = 9,
       height = 1.5)
```


#Main, Downshift - Upshift

```{r}
#Calculate Downshift & Upshift for individual speakers 

temp2 <- aggregate(data1$F1_Change, 
                   list(data1$Direction, data1$Chunk, data1$Speaker), 
                   FUN = mean) 

colnames(temp2) <- c("Direction", "Chunk", "Speaker", "meanF1")

temp2 <- temp2[which(temp2$Direction != "Unshifted"), ]

temp3 <- spread(data = temp2, key = Direction, value = meanF1)

temp3$DownUp <- temp3$Downshift - temp3$Upshift

#Summary

temp2 <- summarySEwithin(data = temp3, 
                         measurevar = "DownUp",
                         withinvars = c("Chunk"), 
                         idvar = "Speaker",
                         na.rm = TRUE)

temp2$Chunk <- as.numeric(temp2$Chunk) * 10
```

```{r}
#Plot
ggplot(data = temp2, aes(x = Chunk, y = DownUp)) +
  geom_point() + 
  geom_line() + 
  geom_errorbar(aes(ymin = DownUp - se, ymax = DownUp + se), 
                width = 4,
                position = position_dodge(0)) + 
  theme_classic() + 
  geom_hline(yintercept = 0, linetype = "solid", 
             color = "black", linewidth = 0.5) + 
  theme() + 
  xlab("Blocks") +
  ylab("Down - Up (mel)") + 
  annotate("rect", xmin = 155-1, xmax = 165-1, ymin = -12, ymax = 75,
           alpha = 0.2,fill = "yellow") + 
  annotate("rect", xmin = 125, xmax = 135, ymin = -12, ymax = 75,
           alpha = 0.2,fill = "yellow") + 
  geom_hline(yintercept = 0, linetype = "solid", 
             color = "black", linewidth = 0.5) + 
  geom_vline(xintercept = bar_1, linetype = "twodash", 
             color = "black", linewidth = 0.7) + 
  geom_vline(xintercept = bar_2, linetype = "twodash", 
             color = "black", linewidth = 0.7) + 
  geom_vline(xintercept = bar_3 -1, linetype = "twodash", 
             color = "black", linewidth = 0.7) + 
  scale_x_continuous(expand = c(0, 0), limits = c(1, 183), 
                     breaks = c(10, 30, 40, 60, 
                                70, 150, 160, 180)) +
  scale_y_continuous(expand = c(0, 0), limits = c(-12, 75), 
                     breaks = c(-10, 0, 70),
                     labels = c("-10", "0", "70"))

ggsave(filename = "H:/Shared drives/SMNG/Experiments/simonMultisyllable/Manuscript/PlotGen/1C.png",
       plot = last_plot(), 
       dpi = 600,
       width = 9,
       height = 2)
```

#Main, Individual participant data

```{r}
temp1 <- summarySEwithin(data = data1, 
                         measurevar = "F1_Change",
                         withinvars = c("Direction", "Chunk", "Speaker"), 
                         na.rm = TRUE)

#Restricted to last 10 trials of Hold and Washout

temp2 <- temp1[which(temp1$Chunk %in% c(13, 16)), ]
temp2 <- temp2[which(temp2$Direction != "Unshifted"), ]

temp2$Chunk <- ifelse(temp2$Chunk == 13, "Hold", "Washout")

temp2$Direction <- factor(temp2$Direction, levels = c("Downshift", "Upshift"))
```

```{r}
ggplot(data = temp2, aes(x = Direction, y = F1_Change, 
                               group = Speaker, color = Speaker)) +
  geom_line() + 
  geom_point() + 
  theme_classic() + 
  ylab("F1 change (mel)") + 
  geom_hline(yintercept = 0, linetype = "dashed", 
             color = "black", linewidth = 0.5) + 
  theme(legend.position = "none",
        axis.title.x = element_blank(),
        #axis.text = element_blank(),
        strip.text.x = element_text(size = 15)) +
  facet_grid(~ Chunk) + 
  scale_y_continuous(expand = c(0, 0), limits = c(-105, 125), 
                     breaks = seq(-100, 100, by = 50))

ggsave(filename = "H:/Shared drives/SMNG/Experiments/simonMultisyllable/Manuscript/PlotGen/1D.png",
       plot = last_plot(), 
       dpi = 600,
       width = 9,
       height = 4)
```

#Save

```{r}
save.image("Exp1.RData")
```
